chore: rename to OpenFileDialog
