+++ src/components/OpenFileDialog/VisTypeFilter.js
